+++ tests/test_fsevents.py
@@ -10,9 +10,7 @@ if not platform.is_darwin():  # noqa
 import logging
 import os
 import time
-from functools import partial
 from os import mkdir, rmdir
-from queue import Queue
 from random import random
 from threading import Thread
 from time import sleep
@@ -25,36 +23,11 @@ from watchdog.observers import Observer
 from watchdog.observers.api import ObservedWatch
 from watchdog.observers.fsevents import FSEventsEmitter
 
-from .shell import mkdtemp, rm, touch
+from .shell import touch
+from .util import TestEventQueue, P, StartWatching
 
 logging.basicConfig(level=logging.DEBUG)
 logger = logging.getLogger(__name__)
-
-
-def setup_function(function):
-    global p, event_queue
-    tmpdir = os.path.realpath(mkdtemp())
-    p = partial(os.path.join, tmpdir)
-    event_queue = Queue()
-
-
-def teardown_function(function):
-    try:
-        emitter.stop()
-        emitter.join(5)
-        assert not emitter.is_alive()
-    except NameError:
-        pass  # `name 'emitter' is not defined` unless we call `start_watching`
-    rm(p(""), recursive=True)
-
-
-def start_watching(path=None, recursive=True, use_full_emitter=False):
-    global emitter
-    path = p("") if path is None else path
-    emitter = FSEventsEmitter(
-        event_queue, ObservedWatch(path, recursive=recursive), suppress_history=True
-    )
-    emitter.start()
 
 
 @pytest.fixture
@@ -100,7 +73,7 @@ def test_coalesced_event_check(event, expectation):
     assert event.is_coalesced == expectation
 
 
-def test_add_watch_twice(observer):
+def test_add_watch_twice(observer: Observer, p: P):
     """Adding the same watch twice used to result in a null pointer return without an exception.
 
     See https://github.com/gorakhargosh/watchdog/issues/765
@@ -121,7 +94,12 @@ def test_add_watch_twice(observer):
     rmdir(a)
 
 
-def test_watcher_deletion_while_receiving_events_1(caplog, observer):
+def test_watcher_deletion_while_receiving_events_1(
+    caplog,
+    p: P,
+    emitter: FSEventsEmitter,
+    start_watching: StartWatching,
+) -> None:
     """
     When the watcher is stopped while there are events, such exception could happen:
 
@@ -153,7 +131,12 @@ def test_watcher_deletion_while_receiving_events_1(caplog, observer):
         assert not caplog.records
 
 
-def test_watcher_deletion_while_receiving_events_2(caplog):
+def test_watcher_deletion_while_receiving_events_2(
+    caplog,
+    p: P,
+    start_watching: StartWatching,
+    emitter: FSEventsEmitter,
+) -> None:
     """Note: that test takes about 20 seconds to complete.
 
     Quite similar test to prevent another issue
@@ -204,7 +187,7 @@ def test_watcher_deletion_while_receiving_events_2(caplog):
         assert not caplog.records
 
 
-def test_remove_watch_twice():
+def test_remove_watch_twice(start_watching: StartWatching, emitter: FSEventsEmitter) -> None:
     """
     ValueError: PyCapsule_GetPointer called with invalid PyCapsule object
     The above exception was the direct cause of the following exception:
@@ -228,7 +211,7 @@ def test_remove_watch_twice():
     emitter.stop()
 
 
-def test_unschedule_removed_folder(observer):
+def test_unschedule_removed_folder(observer, p: P) -> None:
     """
     TypeError: PyCObject_AsVoidPtr called with null pointer
     The above exception was the direct cause of the following exception:
@@ -250,7 +233,12 @@ def test_unschedule_removed_folder(observer):
     observer.unschedule(w)
 
 
-def test_converting_cfstring_to_pyunicode():
+def test_converting_cfstring_to_pyunicode(
+    p: P,
+    start_watching: StartWatching,
+    emitter: FSEventsEmitter,
+    event_queue: TestEventQueue,
+) -> None:
     """See https://github.com/gorakhargosh/watchdog/issues/762"""
 
     tmpdir = p()
@@ -266,7 +254,7 @@ def test_converting_cfstring_to_pyunicode():
         emitter.stop()
 
 
-def test_recursive_check_accepts_relative_paths():
+def test_recursive_check_accepts_relative_paths(p: P) -> None:
     """See https://github.com/gorakhargosh/watchdog/issues/797
 
     The test code provided in the defect observes the current working directory
@@ -316,7 +304,7 @@ def test_recursive_check_accepts_relative_paths():
         observer.join()
 
 
-def test_watchdog_recursive():
+def test_watchdog_recursive(p: P) -> None:
     """See https://github.com/gorakhargosh/watchdog/issues/706"""
     import os.path
