+++ tests/test_inotify_c.py
@@ -7,50 +7,20 @@ from watchdog.utils import platform
 if not platform.is_linux():  # noqa
     pytest.skip("GNU/Linux only.", allow_module_level=True)
 
-import contextlib
 import ctypes
 import errno
 import logging
 import os
 import struct
-from functools import partial
-from queue import Queue
 from unittest.mock import patch
 
 from watchdog.events import DirCreatedEvent, DirDeletedEvent, DirModifiedEvent
-from watchdog.observers.api import ObservedWatch
-from watchdog.observers.inotify import InotifyEmitter, InotifyFullEmitter
 from watchdog.observers.inotify_c import Inotify, InotifyConstants, InotifyEvent
 
-from .shell import mkdtemp, rm
+from .util import TestEventQueue, P, StartWatching
 
 logging.basicConfig(level=logging.DEBUG)
 logger = logging.getLogger(__name__)
-
-
-def setup_function(function):
-    global p, event_queue
-    tmpdir = os.path.realpath(mkdtemp())
-    p = partial(os.path.join, tmpdir)
-    event_queue = Queue()
-
-
-@contextlib.contextmanager
-def watching(path=None, use_full_emitter=False):
-    path = p("") if path is None else path
-    global emitter
-    Emitter = InotifyFullEmitter if use_full_emitter else InotifyEmitter
-    emitter = Emitter(event_queue, ObservedWatch(path, recursive=True))
-    emitter.start()
-    yield
-    emitter.stop()
-    emitter.join(5)
-
-
-def teardown_function(function):
-    rm(p(""), recursive=True)
-    with contextlib.suppress(NameError):
-        assert not emitter.is_alive()
 
 
 def struct_inotify(wd, mask, cookie=0, length=0, name=b""):
@@ -66,7 +36,7 @@ def struct_inotify(wd, mask, cookie=0, length=0, name=b""):
     return struct.pack(struct_format, wd, mask, cookie, length, name)
 
 
-def test_late_double_deletion():
+def test_late_double_deletion(helper: Helper, p: P, event_queue: TestEventQueue, start_watching: StartWatching):
     inotify_fd = type("FD", (object,), {})()
     inotify_fd.last = 0
     inotify_fd.wds = []
@@ -125,7 +95,8 @@ def test_late_double_deletion():
     mock4 = patch.object(inotify_c, "inotify_add_watch", new=inotify_add_watch)
     mock5 = patch.object(inotify_c, "inotify_rm_watch", new=inotify_rm_watch)
 
-    with mock1, mock2, mock3, mock4, mock5, watching(p("")):
+    with mock1, mock2, mock3, mock4, mock5:
+        start_watching(p(""))
         # Watchdog Events
         for evt_cls in [DirCreatedEvent, DirDeletedEvent] * 2:
             event = event_queue.get(timeout=5)[0]
@@ -134,6 +105,7 @@ def test_late_double_deletion():
             event = event_queue.get(timeout=5)[0]
             assert isinstance(event, DirModifiedEvent)
             assert event.src_path == p("").rstrip(os.path.sep)
+        helper.close()
 
     assert inotify_fd.last == 3  # Number of directories
     assert inotify_fd.buf == b""  # Didn't miss any event
@@ -159,31 +131,31 @@ def test_raise_error(error, patterns):
     assert any(pattern in str(exc.value) for pattern in patterns)
 
 
-def test_non_ascii_path():
+def test_non_ascii_path(p: P, event_queue: TestEventQueue, start_watching: StartWatching):
     """
     Inotify can construct an event for a path containing non-ASCII.
     """
     path = p("\N{SNOWMAN}")
-    with watching(p("")):
-        os.mkdir(path)
-        event, _ = event_queue.get(timeout=5)
-        assert isinstance(event.src_path, type(""))
-        assert event.src_path == path
-        # Just make sure it doesn't raise an exception.
-        assert repr(event)
+    start_watching(p(""))
+    os.mkdir(path)
+    event, _ = event_queue.get(timeout=5)
+    assert isinstance(event.src_path, type(""))
+    assert event.src_path == path
+    # Just make sure it doesn't raise an exception.
+    assert repr(event)
 
 
-def test_watch_file():
+def test_watch_file(p: P, event_queue: TestEventQueue, start_watching: StartWatching):
     path = p("this_is_a_file")
     with open(path, "a"):
         pass
-    with watching(path):
-        os.remove(path)
-        event, _ = event_queue.get(timeout=5)
-        assert repr(event)
+    start_watching(path)
+    os.remove(path)
+    event, _ = event_queue.get(timeout=5)
+    assert repr(event)
 
 
-def test_event_equality():
+def test_event_equality(p: P):
     wd_parent_dir = 42
     filename = "file.ext"
     full_path = p(filename)
